+++ example/lib/home_page.dart
@@ -20,6 +20,7 @@ import 'showcases/discovery_bar_showcase.dart';
 import 'showcases/labeled_progress_indicator_showcase.dart';
 import 'showcases/quick_switcher_showcase.dart';
 import 'showcases/stacked_toast_showcase.dart';
+import 'showcases/disclosure_switch_showcase.dart';
 
 
 
@@ -250,6 +251,12 @@ class _HomePageState extends State<HomePage> {
         icon: Icons.notification_important_outlined,
         category: 'Interactions',
         page: StackedToastShowcase(),
+      ),
+      _ComponentItem(
+        title: 'Disclosure Sw',
+        icon: Icons.toggle_on_outlined,
+        category: 'Interactions',
+        page: DisclosureSwitchShowcase(),
       ),
     ];